requirements.txt
tensorflow==2.16.1
tensorflow-text==2.16.1
Flask==3.0.3
google-cloud-storage
